added action to compare button
